% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{audio_file_melodic_production_block}
\alias{audio_file_melodic_production_block}
\title{Audio melodic production block}
\usage{
audio_file_melodic_production_block(
  audio_df,
  max_goes = 3L,
  user_id = 1L,
  paradigm_type = c("call_and_response", "simultaneous_recall")
)
}
\arguments{
\item{paradigm_type}{}
}
\description{
Audio melodic production block
}
