% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{calc.breadth}
\alias{calc.breadth}
\title{calc.breadth}
\usage{
calc.breadth(bed)
}
\arguments{
\item{bed}{imported bed data frame from bedpe object}
}
\value{
the breadth of the bed file
}
\description{
calc.breadth
}
