% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{calc_madOffset}
\alias{calc_madOffset}
\title{calc_madOffset}
\usage{
calc_madOffset(values)
}
\arguments{
\item{values}{vector of values to calculate the MAD offset}
}
\value{
MAD offset
}
\description{
calc_madOffset
}
