% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{clustering}
\usage{
clustering(knn_matrix, feat_matrix, method)
}
\arguments{
\item{knn_matrix}{a matrix of k-nearest neighbors}

\item{feat_matrix}{a matrix of features for each cell}

\item{method}{a function to calculate the affinity between cells}
}
\value{
a vector of cluster membership for each cell
}
\description{
clustering
}
