% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{convert_long}
\alias{convert_long}
\title{convert_long}
\usage{
convert_long(reads)
}
\arguments{
\item{reads}{reads object from QDNASeq}
}
\value{
table of reads for each cell
}
\description{
convert_long
}
