% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{convert_mtxlong}
\alias{convert_mtxlong}
\title{convert_mtxlong}
\usage{
convert_mtxlong(mtx)
}
\arguments{
\item{mtx}{a matrix to convert to long format}
}
\value{
a long format data frame
}
\description{
convert_mtxlong
}
