% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNCalling.R
\name{copyCall}
\alias{copyCall}
\title{copyCall}
\usage{
copyCall(sbird_sce, num_cores = NULL)
}
\arguments{
\item{sbird_sce}{the songbird single cell experiment object - requires that 'segmented' and 'reads' assays are present}

\item{num_cores}{number of cores to use for parallel processing, default is all but one core}

\item{tune_uniploid}{whether to tune the uniploid value for each bin, default is FALSE}
}
\value{
songbird object with the copy number called for each bin
}
\description{
copyCall
}
