% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{correct_ploidy}
\alias{correct_ploidy}
\title{Title}
\usage{
correct_ploidy(ploidy, prop_doublets, coverage)
}
\arguments{
\item{ploidy}{estimated ploidy}

\item{prop_doublets}{proportion of doublets}

\item{coverage}{coverage of the cell}
}
\description{
Title
}
