% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_music_test_app.R
\name{create_music_test_app}
\alias{create_music_test_app}
\title{Create music test app}
\usage{
create_music_test_app(force_p_id_from_url = FALSE, pre_post = c("pre", "post"))
}
\arguments{
\item{pre_post}{}
}
\description{
Create music test app
}
