% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_questionnaire_app.R
\name{create_questionnaire_app}
\alias{create_questionnaire_app}
\title{Create questionnaire app}
\usage{
create_questionnaire_app(
  force_p_id_from_url = FALSE,
  type = c("kids", "parents", "teachers"),
  pre_post = c("pre", "post"),
  img_dir = system.file("data-raw/questionnaires/www/img", package = "songbird"),
  extra_materials = NULL
)
}
\arguments{
\item{extra_materials}{}
}
\description{
Create questionnaire app
}
