% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNCalling.R
\name{create_sce}
\alias{create_sce}
\title{create_sce}
\usage{
create_sce(res)
}
\arguments{
\item{res}{a list of results from cell processing}
}
\value{
the initial songbird object
}
\description{
create_sce
}
