% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNCalling.R
\name{create_sce_from_res}
\alias{create_sce_from_res}
\title{create_sce_from_res}
\usage{
create_sce_from_res(res, n_cpu = NULL)
}
\arguments{
\item{res}{a list of results similar to the cell processing results}

\item{n_cpu}{number of cores to use for parallel processing, default is all but one core}
}
\value{
songbird object with the reads, counts, and segmented data
}
\description{
create_sce_from_res
}
