% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_questionnaires.R
\name{create_timeline}
\alias{create_timeline}
\title{Create timeline}
\usage{
create_timeline(
  type = c("kids", "teachers", "parents"),
  pre_post = c("pre", "post")
)
}
\arguments{
\item{pre_post}{}
}
\description{
Create timeline
}
