% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNCalling.R
\name{detect_wgd}
\alias{detect_wgd}
\title{detect_wgd}
\usage{
detect_wgd(high_qPloidies, all_ploidies)
}
\arguments{
\item{high_qPloidies}{vector of ploidies that are used for correction}

\item{all_ploidies}{all ploidies}
}
\value{
the WGD identity of cells - only detects 1 whole genome duplication
}
\description{
detect_wgd
}
