% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{extractStrandedness}
\alias{extractStrandedness}
\title{extractStrandedness}
\usage{
extractStrandedness(readName, n)
}
\arguments{
\item{readName}{name for each read pulled from the bed file}

\item{n}{number of characters to count back to get strandedness}
}
\value{
read orientation from the last n characters of the read name
}
\description{
extractStrandedness
}
