% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{filter.bed}
\alias{filter.bed}
\title{filter.bed}
\usage{
filter.bed(bed, bin_data)
}
\arguments{
\item{bed}{a data frame formatted as a bed file with read positions}

\item{bin_data}{the metadata generated by QDNAseq for the bins}
}
\value{
a data frame with duplicate reads (identical start or end sites) removed
}
\description{
filter.bed
}
