% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNCalling.R
\name{fitMeans}
\alias{fitMeans}
\title{fitMeans}
\usage{
fitMeans(means, use, sigma, expected_ploidy = NA)
}
\arguments{
\item{means}{numeric vector of segmented read counts to fit}

\item{use}{logical vector indicating which bins are good to use}

\item{sigma}{estimate of the noise within the bins - calculated by the residual of the reads - segmented state}

\item{expected_ploidy}{noisy estimator of true cell ploidy - NA results in testing all Cell Ploidies 2-8}
}
\value{
The final CN state for each bin
}
\description{
fitMeans
}
