% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_CNCalling.R
\name{gauss_kernel}
\alias{gauss_kernel}
\title{Title}
\usage{
gauss_kernel(matrix, n_neighbors)
}
\arguments{
\item{matrix}{the inputted matrix to smooth over}

\item{n_neighbors}{number of adjacent bins to smooth with}
}
\description{
Title
}
