% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{gen_annobar}
\alias{gen_annobar}
\title{gen_annobar}
\usage{
gen_annobar(values, orientation, ylim = NULL)
}
\arguments{
\item{values}{a vector of values to plot as an annotation bar}

\item{orientation}{a string indicating the orientation of the annotation bar, either 'row' or 'column'}

\item{ylim}{a numeric vector of length 2 indicating the limits of the y-axis for the barplot. If NULL, the limits will be set to the 1st and 99th percentiles of the values.}
}
\value{
a ComplexHeatmap annotation object
}
\description{
gen_annobar
}
