% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_CNCalling.R
\name{generate_changepoint_matrix}
\alias{generate_changepoint_matrix}
\title{Title}
\usage{
generate_changepoint_matrix(matrix, use_mask)
}
\arguments{
\item{matrix}{the inputted matrix to identify changepoints from}

\item{use_mask}{bins to use}
}
\description{
Title
}
