% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{get_binMetadata}
\alias{get_binMetadata}
\title{get_binMetadata}
\usage{
get_binMetadata(sce)
}
\arguments{
\item{sce}{songbird object}
}
\value{
column metadata for the songbird object
}
\description{
get_binMetadata
}
