% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{identify_subclones}
\alias{identify_subclones}
\title{Identify Subclones using breakpoint identification and leiden clustering. Clusters with fewer than min_size cells are marked as noisy and put in cluster 0.}
\usage{
identify_subclones(
  sbird_sce,
  assay = "segmented",
  k = 30,
  method = inv_manhattan,
  min_size = 5,
  seed = 1234,
  min_readCount = 1e+05,
  column_name = "subclone"
)
}
\arguments{
\item{sbird_sce}{songbird object}

\item{assay}{the assay to use for clustering}

\item{k}{number of nearest neighbors to use for clustering}

\item{method}{the clustering method to use, default is inverse manhattan distance}

\item{min_size}{minimum size of a subclone to be considered, default is 5 cells}

\item{seed}{random seed for reproducibility}

\item{min_readCount}{minimum read count for a cell to be considered, default is 1e5}

\item{column_name}{the column name to store the subclone information in, default is 'subclone'}
}
\value{
A SingleCellExperiment object with clonal membership
}
\description{
Identify Subclones using breakpoint identification and leiden clustering. Clusters with fewer than min_size cells are marked as noisy and put in cluster 0.
}
