% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{inv_manhattan}
\alias{inv_manhattan}
\title{inv_manhattan}
\usage{
inv_manhattan(x, y)
}
\arguments{
\item{matrix}{the inputted matrix to identify changepoints from}

\item{use_mask}{bins to use}
}
\value{
a matrix of changepoints
}
\description{
inv_manhattan
}
