% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{inv_wasserstein}
\alias{inv_wasserstein}
\title{inv_wasserstein}
\usage{
inv_wasserstein(x, y)
}
\arguments{
\item{x}{a vector of changepoints}

\item{y}{a vector of changepoints}
}
\value{
the inverse Wasserstein distance between the two changepoints
}
\description{
inv_wasserstein
}
