% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_audio_item_receiver.R
\name{launch_audio_item_receiver}
\alias{launch_audio_item_receiver}
\title{Launch audio item receiver}
\usage{
launch_audio_item_receiver(
  max_goes = 3L,
  paradigm_type = c("call_and_response", "simultaneous_recall"),
  app_name,
  app_url = "https://dev.singpause.songbird.training/dashboard/"
)
}
\arguments{
\item{app_url}{}
}
\description{
Launch audio item receiver
}
