% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{load_cell}
\alias{load_cell}
\title{load_cell}
\usage{
load_cell(bamPath, binSize, genome)
}
\arguments{
\item{bamPath}{path to the bam file}

\item{binSize}{size of the bins in # of nucleotides}

\item{genome}{genome to use for binning}
}
\description{
load_cell
}
