% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNCalling.R
\name{ploidy_correction}
\alias{ploidy_correction}
\title{ploidy_correction}
\usage{
ploidy_correction(sbird_sce, min_reads = 1e+05, column = "subclone")
}
\arguments{
\item{sbird_sce}{the songbird object after subclone calling}

\item{min_reads}{minimum number of reads to use a cell's ploidy estimation for correction}

\item{column}{the column in colData to use for subclone membership, default is 'subclone'}
}
\value{
songbird object with corrected ploidy estimates
}
\description{
ploidy_correction
}
