% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_cell}
\alias{plot_cell}
\title{plot_cell}
\usage{
plot_cell(sbird_sce, cell, assay = "copy", chr = NULL, return_plot = FALSE)
}
\arguments{
\item{sbird_sce}{songbird object}

\item{cell}{cell name to plot}

\item{assay}{assay to plot. Will be overlaid over reads}

\item{chr}{chromosome to plot, if NULL all chromosomes will be plotted}

\item{return_plot}{if TRUE, returns the plot object instead of plotting it}
}
\value{
a ggplot object or plots the data
}
\description{
plot_cell
}
