% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Data Heatmap Plotter}
\usage{
plot_heatmap(
  sce,
  assay_name,
  cell_attribs = NULL,
  row_split = NULL,
  rowClust = T,
  use_raster = TRUE,
  bin_attribs = NULL,
  return = FALSE,
  row_title = NULL,
  plot_title = NULL,
  legend_title = NULL,
  show_cellNames = FALSE
)
}
\arguments{
\item{sce}{songbird object}

\item{assay_name}{the name of the assay to plot}

\item{cell_attribs}{column name of the metadata you want plotted next to the cells}

\item{row_split}{vector of identities to group the cells - typically subclonal id}

\item{rowClust}{boolean indicating whether to cluster the rows of the heatmap}

\item{use_raster}{boolean indicating whether to use raster graphics for the heatmap}

\item{bin_attribs}{column name of the column attributes to plot above the heatmap}

\item{return}{boolean indicating whether to return the heatmap object or just plot it out}

\item{row_title}{title for the row labels of the heatmap}

\item{plot_title}{title for the heatmap}

\item{legend_title}{title for the heatmap legend}

\item{show_cellNames}{boolean indicating whether to show the cell names on the heatmap}
}
\value{
a heatmap object if return = TRUE, otherwise plots the heatmap
}
\description{
Data Heatmap Plotter
}
