% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{process.bam}
\alias{process.bam}
\title{process.bam}
\usage{
process.bam(
  bam,
  genome,
  bin.size = 5e+05,
  min_length = 50,
  max_length = NULL,
  tag_overlap = 10,
  focal_amps = TRUE
)
}
\arguments{
\item{bam}{path to a bam file}

\item{genome}{genome assembly to use (e.g. hg38, hg19)}

\item{bin.size}{number of nucleotides per bin}

\item{min_length}{minimum length for reads}

\item{max_length}{maximum length for reads}

\item{tag_overlap}{size of the read overlap created by a tagmentation event}

\item{bedpe}{path to the accompanying bedfile}
}
\value{
corrected reads
}
\description{
process.bam
}
