% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{prune_offsets}
\alias{prune_offsets}
\title{prune_offsets}
\usage{
prune_offsets(ubh_obj, reads, min_svSize)
}
\arguments{
\item{ubh_obj}{path to the bedpe file produced by the preprocessing pipeline}

\item{reads}{bin data from the QDNASeq object}

\item{min_svSize}{minimum read length (default 30 nt)}
}
\value{
a data frame set up as a traditional bed file
}
\description{
prune_offsets
}
