% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNCalling.R
\name{reads_to_matrix}
\alias{reads_to_matrix}
\title{Reads_to_Matrix
This function converts a long form data into a matrix}
\usage{
reads_to_matrix(data, column, return_counts = FALSE)
}
\arguments{
\item{data}{list of cell processing results}

\item{column}{column to select from}

\item{return_counts}{boolean to return total reads}
}
\value{
a matrix of reads or counts
}
\description{
Reads_to_Matrix
This function converts a long form data into a matrix
}
