% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{remove.duplicates}
\alias{remove.duplicates}
\title{Title}
\usage{
remove.duplicates(bed)
}
\arguments{
\item{bed}{a data frame formatted as a bed file with read positions}
}
\value{
a data frame with duplicate reads not caught by samtools removed
}
\description{
Title
}
