% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNCalling.R
\name{sanitize_ploidy}
\alias{sanitize_ploidy}
\title{Sanitize_Ploidy}
\usage{
sanitize_ploidy(expected_ploidy)
}
\arguments{
\item{expected_ploidy}{noisy estimator of true cell ploidy}
}
\value{
a vector of possible modal ploidy for the cell
}
\description{
Sanitize_Ploidy
}
