% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNCalling.R
\name{tune_rpcn}
\alias{tune_rpcn}
\title{tune_rpcn}
\usage{
tune_rpcn(values, null, uniploid, sigma)
}
\arguments{
\item{values}{numeric vector of segmented read counts to fit}

\item{null}{numeric vector of null values to compare against}

\item{uniploid}{estimated ratio of reads per copy number}

\item{sigma}{estimate of the noise within the bins}
}
\value{
negative log likelihood (for minimization by an optimizer)
}
\description{
tune_rpcn
}
