% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{ubh_segment}
\alias{ubh_segment}
\title{ubh_segment}
\usage{
ubh_segment(reads, use, min_svSize)
}
\arguments{
\item{reads}{gc and map corrected reads from the qDNASeq object}

\item{use}{boolean array marking high quality bins}

\item{min_svSize}{minimum confident structural variant size (in bins)}
}
\description{
ubh_segment
}
