% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{wh_transform}
\alias{wh_transform}
\title{wh_transform}
\usage{
wh_transform(vals, sv_binSize = 25)
}
\arguments{
\item{vals}{values for the walsh haddamayer transform}

\item{sv_binSize}{minimum confident structural variant size}
}
\description{
wh_transform
}
