% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNCalling.R
\name{copyCall}
\alias{copyCall}
\title{copyCall}
\usage{
copyCall(sbird_sce, num_cores = NULL, tune_uniploid = FALSE)
}
\arguments{
\item{sbird_sce}{the songbird single cell experiment object - requires that 'segmented' and 'reads' assays are present}
}
\value{
songbird object with the copy number called for each bin
}
\description{
copyCall
}
