% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{count.doublets}
\alias{count.doublets}
\title{Title}
\usage{
count.doublets(bed, tag.overlap = 9)
}
\arguments{
\item{bed}{a data frame formatted as a bed file with read positions must be filtered by remove.duplicates}

\item{tag.overlap}{smallest length of the tagmentation read overlap (default 9 nt)}
}
\description{
Title
}
