% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{count.overlaps}
\alias{count.overlaps}
\title{count.overlaps}
\usage{
count.overlaps(bed, min.size = 50, max.size = 1000, tag.overlap = 10)
}
\arguments{
\item{bed}{a data frame formatted as a bed file with read positions must be filtered by remove.duplicates}

\item{min.size}{minimum read length (default 50 nt)}

\item{max.size}{maximum read length (default 1000 nt)}

\item{tag.overlap}{largest length of the tagmentation read overlap (default 10 nt)}
}
\value{
the bed file with the number of reads overlapping the upstream and over regions for each read
}
\description{
count.overlaps
}
