% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{estimate.ploidy}
\alias{estimate.ploidy}
\title{Title}
\usage{
estimate.ploidy(sample, binSize, min_length = 50, max_length = 1000)
}
\arguments{
\item{sample}{path to bedpe file}

\item{binSize}{binning size in nucleotides}

\item{min_length}{minimum read length (default 50 nt)}

\item{max_length}{maximum read length (default 1000 nt)}
}
\description{
Title
}
