% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_CNCalling.R
\name{fitMeans}
\alias{fitMeans}
\title{Title}
\usage{
fitMeans(means, use, expected_ploidy, sigma = 0.5)
}
\arguments{
\item{means}{output from the unbalanced haar transform}

\item{use}{boolean array marking high quality bins}

\item{expected_ploidy}{noisy estimator of true cell ploidy}

\item{sigma}{standard deviation of the normal distribution}
}
\description{
Title
}
