% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{generate_changepoint_matrix}
\alias{generate_changepoint_matrix}
\title{generate_changepoint_matrix}
\usage{
generate_changepoint_matrix(matrix, bin_mask, cell_mask)
}
\arguments{
\item{matrix}{the inputted matrix to identify changepoints from}

\item{use_mask}{bins to use}
}
\value{
a matrix of changepoints
}
\description{
generate_changepoint_matrix
}
