% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{inv_manhattan}
\alias{inv_manhattan}
\title{inv_manhattan}
\usage{
inv_manhattan(x, y)
}
\arguments{
\item{x}{a vector of changepoints}

\item{y}{a vector of changepoints}
}
\value{
the inverse distance between the two changepoints
}
\description{
inv_manhattan
}
