% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{is.doublet}
\alias{is.doublet}
\title{count.doublets}
\usage{
is.doublet(bed, min.tag.overlap = 9, max.tag.overlap = 10)
}
\arguments{
\item{bed}{a data frame formatted as a bed file with read positions must be filtered by filter.bed}

\item{min.tag.overlap}{smallest length of the tagmentation read overlap (default 9 nt)}

\item{max.tag.overlap}{largest length of the tagmentation read overlap (default 10 nt)}
}
\value{
a boolean for each read if it is a doublet, i.e. comes from 2 successful sequential tagmentation events
}
\description{
count.doublets
}
