% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{load.preprocess.bed}
\alias{load.preprocess.bed}
\title{load.preprocess.bed}
\usage{
load.preprocess.bed(bedpe_file, min_length = 30, max_length = 1000)
}
\arguments{
\item{bedpe_file}{path to the bedpe file produced by the preprocessing pipeline}

\item{min_length}{minimum read length (default 30 nt)}

\item{max_length}{maximum read length (default 1000 nt)}
}
\value{
a data frame set up as a traditional bed file
}
\description{
load.preprocess.bed
}
