% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{process.batch}
\alias{process.batch}
\title{process.batch}
\usage{
process.batch(
  bams,
  genome = "hg38",
  bedpes = NULL,
  bin.size = 5e+05,
  min_length = 50,
  max_length = 1000,
  tag_overlap = 9,
  n_cpu = NULL
)
}
\arguments{
\item{bams}{path to a bam file}

\item{genome}{genome assembly to use (e.g. hg38, hg19)}

\item{bedpes}{path to the accompanying bedfile}

\item{bin.size}{number of nucleotides per bin}

\item{min_length}{minimum length for reads}

\item{max_length}{maximum length for reads}

\item{tag_overlap}{size of the read overlap created by a tagmentation event}

\item{n_cpu}{number of CPUs to use for parallel processing, default is NULL which uses all but one CPU}
}
\value{
songbird object
}
\description{
process.batch
}
