% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{process.batch}
\alias{process.batch}
\title{process.batch}
\usage{
process.batch(
  bams,
  genome = "hg38",
  bedpes = NULL,
  bin.size = 5e+05,
  min.svSize = 1e+06,
  min_length = 50,
  max_length = 1000,
  tag_overlap = 9,
  ext_correction = NULL,
  n_cpu = NULL
)
}
\arguments{
\item{bams}{path to a bam file}

\item{genome}{estimated ploidy of the cell (from estimate.ploidy)}

\item{bedpes}{path to the accompanying bedfile}

\item{bin.size}{number of nucleotides per bin}

\item{min.length}{number of bins for the smallest confident structural variant}
}
\value{
songbird object
}
\description{
process.batch
}
