% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{process.cell}
\alias{process.cell}
\title{process.cell}
\usage{
process.cell(
  bam,
  genome,
  bedpe = NULL,
  bin.size = 5e+05,
  min_length = 50,
  max_length = 1000,
  tag_overlap = 9
)
}
\arguments{
\item{bam}{path to a bam file}

\item{genome}{genome assembly to use (e.g. hg38, hg19)}

\item{bedpe}{path to the accompanying bedfile}

\item{bin.size}{number of nucleotides per bin}

\item{min_length}{minimum length for reads}

\item{max_length}{maximum length for reads}

\item{tag_overlap}{size of the read overlap created by a tagmentation event}
}
\value{
corrected reads
}
\description{
process.cell
}
