% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{process.cell}
\alias{process.cell}
\title{process.cell}
\usage{
process.cell(
  bam,
  bedpe = NULL,
  bin.size = 5e+05,
  min.svSize = 1e+06,
  min_length = 50,
  max_length = 1000
)
}
\arguments{
\item{bam}{path to a bam file}

\item{bedpe}{path to the accompanying bedfile}

\item{bin.size}{number of nucleotides per bin}

\item{min.svSize}{number of bins for the smallest confident structural variant}

\item{est.ploidy}{estimated ploidy of the cell (from estimate.ploidy)}
}
\description{
process.cell
}
