% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_CNCalling.R
\name{reads_to_matrix}
\alias{reads_to_matrix}
\title{Title}
\usage{
reads_to_matrix(data, column, return_counts = FALSE)
}
\arguments{
\item{data}{list of cell processing results}

\item{column}{column to select from}

\item{return_counts}{boolean to return total reads}
}
\description{
Title
}
