% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedProcessing.R
\name{ubh_segment}
\alias{ubh_segment}
\title{ubh_segment}
\usage{
ubh_segment(values, use)
}
\arguments{
\item{values}{GC + Map corrected reads per bin}

\item{use}{bins to use}
}
\description{
ubh_segment
}
